Add files via upload
This file performs credit card fraud detection with a stacking ensemble approach. It uses 5 base classifiers which includes k-NN, SVM Classifier, Gaussian Naive Bayes, Decision Tree (CART) and Random Forest Classifier. It uses Logistic Regression as the meta-classifier. It gives 5 performance metrics which include accuracy, precision, recall, f1-score and ROC-AUC. It is possible to fit and predict new data with the model so long as the features in the test data corresponds to the number features in the original training set.